requirement.txt
beautifulsoup4==4.3.2
crispy-semantic-ui==1.0.0
Django==1.8.2
django-activelink==0.4
django-autoslug==1.8.0
django-braces==1.8.0
django-coverage==1.2.4
django-crispy-forms==1.4.0
django-debug-toolbar==1.3.0
django-downloadview==1.6
django-easy-pdf==0.1.0
django-emoticons==1.1.2
django-factory-boy==0.1.6
django-filepicker==0.2.2
django-filer==0.9.11
django-floppyforms==1.4.1
django-markdown==0.8.4
django-model-utils==2.2
django-mptt==0.7.4
django-mptt-admin==0.2.1
django-pagedown==0.1.0
django-polymorphic==0.7.1
django-suit==0.2.13
django-vanilla-views==1.0.4
easy-thumbnails==2.2
ecdsa==0.13
Embedly==0.5.0
Fabric==1.10.1
factory-boy==2.5.2
html5lib==0.99999
httplib2==0.9.1
Markdown==2.6.2
mock==1.0.1
paramiko==1.15.2
Pillow==2.7.0
psycopg2==2.5.5
pycrypto==2.6.1
PyPDF2==1.24
pytz==2015.4
raven==5.3.1
reportlab==3.2.0
requests==2.7.0
six==1.9.0
sqlparse==0.1.15
Unidecode==0.4.17
xhtml2pdf==0.0.6
